+++ buybackprogram/app_settings.py
@@ -11,6 +11,11 @@ EXAMPLE_SETTING_ONE = getattr(settings, "EXAMPLE_SETTING_ONE", None)
 # Hard timeout for tasks in seconds to reduce task accumulation during outages
 BUYBACKPROGRAM_TASKS_TIME_LIMIT = clean_setting("BUYBACKPROGRAM_TASKS_TIME_LIMIT", 7200)
 
+# Warning limit for Jita price updates if prices have not been updated
+BUYBACKPROGRAM_PRICE_AGE_WARNING_LIMIT = clean_setting(
+    "BUYBACKPROGRAM_PRICE_AGE_WARNING_LIMIT", 48
+)
+
 # Tracking number tag
 BUYBACKPROGRAM_TRACKING_PREFILL = clean_setting(
     "BUYBACKPROGRAM_TRACKING_PREFILL", "aa-bbp"